+++ integration/testarchive/archive_test.go
@@ -4,8 +4,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/databricks/cli/internal/testutil"
-	"github.com/databricks/cli/libs/testarchive"
+	"github.com/databricks/cli/internal/testarchive"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
