+++ integration/testarchive/downloader_test.go
@@ -5,8 +5,8 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/databricks/cli/internal/testarchive"
 	"github.com/databricks/cli/internal/testutil"
-	"github.com/databricks/cli/libs/testarchive"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
